# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeetCoach AI API",
    "description": "AI coding mentor for LeetCode \u2014 hints, patterns, complexity, debugging, review, and revision tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ai/hints": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Hints",
        "description": "Progressive hint ladder (never reveals the solution).",
        "operationId": "hints_api_v1_ai_hints_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HintsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HintsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/pattern": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Pattern",
        "description": "Detect the algorithmic pattern and explain why.",
        "operationId": "pattern_api_v1_ai_pattern_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatternResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/complexity": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Complexity",
        "description": "Estimate time/space complexity of the user's code.",
        "operationId": "complexity_api_v1_ai_complexity_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/debug": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Debug",
        "description": "Debug assistant: mistakes, edge cases, missing conditions, tests.",
        "operationId": "debug_api_v1_ai_debug_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebugRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/review": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Review",
        "description": "Code review: naming, readability, optimization, duplication, dead code.",
        "operationId": "review_api_v1_ai_review_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/explain": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Explain",
        "description": "Line-by-line explanation in beginner/intermediate/interview modes.",
        "operationId": "explain_api_v1_ai_explain_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/interview": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Interview",
        "description": "Generate interview follow-up questions.",
        "operationId": "interview_api_v1_ai_interview_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/interview/feedback": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Interview Feedback",
        "description": "Score the user's answer to an interview question.",
        "operationId": "interview_feedback_api_v1_ai_interview_feedback_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewFeedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/solution": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Solution",
        "description": "Explicitly requested solution \u2014 the only endpoint that reveals code.",
        "operationId": "solution_api_v1_ai_solution_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/similar": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Similar",
        "description": "Recommend related problems grouped by difficulty.",
        "operationId": "similar_api_v1_ai_similar_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/daily": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "Daily",
        "description": "Personalized daily challenge.",
        "operationId": "daily_api_v1_ai_daily_get",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyChallenge"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/flashcards/generate": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Generate Flashcards",
        "description": "Generate flashcards from a solved problem (and store them).",
        "operationId": "generate_flashcards_api_v1_ai_flashcards_generate_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashcardGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Generate Flashcards Api V1 Ai Flashcards Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/rag": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "Rag Search",
        "description": "Retrieve DSA knowledge-base chunks (powers grounded chat).",
        "operationId": "rag_search_api_v1_ai_rag_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "K"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rag Search Api V1 Ai Rag Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplexityResponse": {
        "properties": {
          "time_complexity": {
            "type": "string",
            "title": "Time Complexity"
          },
          "space_complexity": {
            "type": "string",
            "title": "Space Complexity"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "optimizations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Optimizations"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "time_complexity",
          "space_complexity",
          "explanation"
        ],
        "title": "ComplexityResponse"
      },
...
      "DebugRequest": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/ProblemContext"
          },
          "error": {
            "type": "string",
            "maxLength": 8000,
            "title": "Error",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "problem"
        ],
        "title": "DebugRequest"
      },
      "DebugResponse": {
        "properties": {
          "possible_mistakes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Possible Mistakes"
          },
          "edge_cases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Edge Cases"
          },
          "missing_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Conditions"
          },
          "suggested_tests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Tests"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "title": "DebugResponse"
      },
      "ExplainRequest": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/ProblemContext"
          },
          "mode": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "interview"
            ],
            "title": "Mode",
            "default": "intermediate"
          }
        },
        "type": "object",
        "required": [
          "problem"
        ],
        "title": "ExplainRequest"
      },
      "ExplainResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "interview"
            ],
            "title": "Mode"
          },
          "overview": {
            "type": "string",
            "title": "Overview"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineExplanation"
            },
            "type": "array",
            "title": "Lines"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "overview"
        ],
        "title": "ExplainResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HintsRequest": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/ProblemContext"
          },
          "levels_to_reveal": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Levels To Reveal",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "problem"
        ],
        "title": "HintsRequest"
      },
      "HintsResponse": {
        "properties": {
          "pattern": {
            "type": "string",
            "title": "Pattern",
            "default": ""
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/HintLevel"
            },
            "type": "array",
            "title": "Levels"
          },
          "code_revealed": {
            "type": "boolean",
            "title": "Code Revealed",
            "default": false
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "title": "HintsResponse"
      },
      "InterviewAnswerRequest": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/ProblemContext"
          },
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "problem",
          "question_id",
          "answer"
        ],
        "title": "InterviewAnswerRequest"
      },
...
      "InterviewResponse": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "title": "InterviewResponse"
      },
...
      "PatternResponse": {
        "properties": {
          "primary": {
            "$ref": "#/components/schemas/PatternMatch"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/PatternMatch"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "when_to_use": {
            "type": "string",
            "title": "When To Use",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "primary"
        ],
        "title": "PatternResponse"
      },
      "ProblemContext": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "default": ""
          },
          "leetcode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leetcode Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "Medium"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "examples": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Examples"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "function_signature": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Function Signature"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "python"
          }
        },
        "type": "object",
        "title": "ProblemContext",
        "description": "Normalised problem payload extracted by the extension's content script."
      },
...
      "ReviewResponse": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            },
            "type": "array",
            "title": "Findings"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "summary"
        ],
        "title": "ReviewResponse"
      },
...
    }
  }
}